Lots of work on improving the floating point emulation. Note that most changes affect only the interpreter mode.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@4314 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Core/Src/PowerPC/Interpreter/Interpreter.h
@@ -307,7 +307,7 @@ namespace Interpreter
 
 	// paired helper
 	float Helper_Dequantize(const u32 _Addr, const EQuantizeType _quantizeType, const unsigned int _uScale);
-	void  Helper_Quantize  (const u32 _Addr, const float _fValue, const EQuantizeType _quantizeType, const unsigned _uScale);
+	void  Helper_Quantize  (const u32 _Addr, const double _fValue, const EQuantizeType _quantizeType, const unsigned _uScale);
 
 	// other helper
 	u32 Helper_Mask(int mb, int me);
